---SwitchWidget---
Created Switch widget to enable and disable the dark and light mode. Used Transform.scale() widget for scaling switch. Text widget for showing current mode of the switch.

autofocus: setting true to focus of the widget is first.
splashRadius: touch/click radius of the thumb.
value: setting bool value to control the switch.
focusNode: Setting this property to focusing the widget.
focusColor: color of the focusing widget.
hoverColor: color of the hover.
activeColor: color of the active switch.
inactiveThumbColor: color of the disabled switch thumb color.
inactiveTrackColor: color of the disabled switch track color.
activeTrackColor: color of the active switch track color.
mouseCursor: setting this to click (Default).
dragStartBehavior: setting start to smooth animation, setting down to flexible animation. Not big difference.
trackOutlineWidth: widget of the switch widget.
materialTapTargetSize: size of the tapping target.
overlayColor: color of the hover.
thumbColor: color of the thumb.
padding: switch padding.
trackOutlineColor: stroke color of the switch.
trackColor: color of the track color.
thumbIcon: separate icons for on and off switch icon of thumb.
onChanged: This method controls the enable/disable of the switch.